+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/role/ProductManager.java
@@ -1,6 +1,6 @@
 package vendingmachine.role;
 
-import vendingmachine.Money;
+import vendingmachine.domain.money.Money;
 import vendingmachine.domain.product.Product;
 
 /*
