+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/view/OutputView.java
@@ -4,8 +4,8 @@ import java.util.Comparator;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
-import vendingmachine.Coin;
-import vendingmachine.Money;
+import vendingmachine.domain.money.Coin;
+import vendingmachine.domain.money.Money;
 
 /**
  * 자판기가 보유하고 있는 금액을 입력해 주세요. 450
